Refactor the initialization file to 'init'
