+++ server/services/discord/actions/on_reaction_add.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const client = require('../connect')
+const client = require('../init')
 
 /**
  * @brief Triggered when a user reacts to a message.
