+++ server/services/discord/reactions/send_private_message.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const client = require('../connect')
+const client = require('../init')
 
 /**
  * @brief Send a private message to a user
